Add Design ALl Page

+++ productos/borrar.php
@@ -11,14 +11,27 @@ if ($conn->connect_error) {
     exit("Conexion fallida: " . $conn->connect_error);
 }
 
-$sql = "DELETE FROM `productos` WHERE `id_producto` = " . $_GET["id"];
 
-if ($conn->query($sql) === TRUE) {
-    header('Location: ./index.php');
-} else {
-    header('Location: ./index.php');
+$sql = "SELECT count(*) FROM facturas_productos WHERE producto_id = ". $_GET["id"];
+$conteo = $conn->query($sql)->fetch_row();
+
+if($conteo[0] > 0){
+	echo '<script>
+			alert("El producto ya se encuentra en una factura");
+ 			window.location="./index.php";
+		  </script>';
+}else{
+
+	$sql = "DELETE FROM `productos` WHERE `id_producto` = " . $_GET["id"];
+
+	if ($conn->query($sql) === TRUE) {
+	    header('Location: ./index.php');
+	} else {
+	    header('Location: ./index.php');
+	}
+
+	$conn->close();
+	die();
+
+
 }
-
-$conn->close();
-
-die();
